+++ app/src/main/java/org/simple/clinic/scanid/QRCodeJsonParser.kt
@@ -1,0 +1,14 @@
+package org.simple.clinic.scanid
+
+import com.squareup.moshi.Moshi
+import javax.inject.Inject
+
+class QRCodeJsonParser @Inject constructor(
+    moshi: Moshi
+) {
+  private val adapter = moshi.adapter(IndiaNHIDInfoPayload::class.java)
+
+  fun parseQRCodeJson(text: String): IndiaNHIDInfoPayload? {
+    return adapter.fromJson(text)
+  }
+}