+++ app/src/main/java/org/simple/clinic/scanid/ScanSimpleIdEffectHandler.kt
@@ -2,7 +2,7 @@ package org.simple.clinic.scanid
 
 import com.spotify.mobius.rx2.RxMobius
 import com.squareup.moshi.JsonDataException
-import com.squareup.moshi.Moshi
+import com.squareup.moshi.JsonEncodingException
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
@@ -14,7 +14,7 @@ import org.simple.clinic.util.scheduler.SchedulersProvider
 class ScanSimpleIdEffectHandler @AssistedInject constructor(
     private val schedulersProvider: SchedulersProvider,
     private val patientRepository: PatientRepository,
-    private val moshi: Moshi,
+    private val qrCodeJsonParser: QRCodeJsonParser,
     private val country: Country,
     @Assisted private val uiActions: ScanSimpleIdUiActions
 ) {
@@ -56,6 +56,8 @@ class ScanSimpleIdEffectHandler @AssistedInject constructor(
               parseJsonBasedOnCountry(it)
             } catch (e: JsonDataException) {
               InvalidQrCode
+            } catch (e: JsonEncodingException) {
+              InvalidQrCode
             }
           }
     }
@@ -64,9 +66,7 @@ class ScanSimpleIdEffectHandler @AssistedInject constructor(
   private fun parseJsonBasedOnCountry(effect: ParseScannedJson): ScanSimpleIdEvent {
     return when (country.isoCountryCode) {
       Country.INDIA -> {
-        val adapter = moshi.adapter(IndiaNHIDInfoPayload::class.java)
-        val payload = adapter.fromJson(effect.text)
-
+        val payload = qrCodeJsonParser.parseQRCodeJson(effect.text)
         ScannedQRCodeJsonParsed(payload?.toPatientPrefillInfo(), payload?.healthIdNumber)
       }
       else -> InvalidQrCode